feat: aucell rust part done
