+++ src/rust/src/single_cell/mod.rs
@@ -1,4 +1,4 @@
-pub mod dge;
+pub mod dge_aucs;
 pub mod fast_ranking;
 pub mod processing;
 pub mod sc_knn_snn;
